+++ spring-boot2/src/main/java/com/apollo/openapi/server/api/PortalManagementApi.java
@@ -8,6 +8,7 @@ package com.apollo.openapi.server.api;
 import com.apollo.openapi.server.model.OpenConsumerCreateRequestDTO;
 import com.apollo.openapi.server.model.OpenConsumerInfoDTO;
 import com.apollo.openapi.server.model.OpenConsumerSummaryDTO;
+import com.apollo.openapi.server.model.OpenConsumerTokenDTO;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -942,7 +943,7 @@ public interface PortalManagementApi {
         tags = { "Portal Management" },
         responses = {
             @ApiResponse(responseCode = "200", description = "成功获取消费者Token", content = {
-                @Content(mediaType = "application/json", schema = @Schema(implementation = OpenConsumerInfoDTO.class))
+                @Content(mediaType = "application/json", schema = @Schema(implementation = OpenConsumerTokenDTO.class))
             })
         },
         security = {
@@ -954,7 +955,7 @@ public interface PortalManagementApi {
         value = "/openapi/v1/consumer-tokens/by-appId",
         produces = { "application/json" }
     )
-    default ResponseEntity<OpenConsumerInfoDTO> getConsumerTokenByAppId(
+    default ResponseEntity<OpenConsumerTokenDTO> getConsumerTokenByAppId(
         @NotNull @Parameter(name = "appId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "appId", required = true) String appId
     ) {
         return getDelegate().getConsumerTokenByAppId(appId);