+++ typescript/src/apis/PortalManagementApi.ts
@@ -17,6 +17,7 @@ import type {
   OpenConsumerCreateRequestDTO,
   OpenConsumerInfoDTO,
   OpenConsumerSummaryDTO,
+  OpenConsumerTokenDTO,
 } from '../models';
 import {
     OpenConsumerCreateRequestDTOFromJSON,
@@ -25,6 +26,8 @@ import {
     OpenConsumerInfoDTOToJSON,
     OpenConsumerSummaryDTOFromJSON,
     OpenConsumerSummaryDTOToJSON,
+    OpenConsumerTokenDTOFromJSON,
+    OpenConsumerTokenDTOToJSON,
 } from '../models';
 
 export interface AddFavoriteRequest {
@@ -1321,7 +1324,7 @@ export class PortalManagementApi extends runtime.BaseAPI {
      * GET /openapi/v1/consumer-tokens/by-appId
      * 按应用ID查询消费者Token(new added)
      */
-    async getConsumerTokenByAppIdRaw(requestParameters: GetConsumerTokenByAppIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<OpenConsumerInfoDTO>> {
+    async getConsumerTokenByAppIdRaw(requestParameters: GetConsumerTokenByAppIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<OpenConsumerTokenDTO>> {
         if (requestParameters.appId === null || requestParameters.appId === undefined) {
             throw new runtime.RequiredError('appId','Required parameter requestParameters.appId was null or undefined when calling getConsumerTokenByAppId.');
         }
@@ -1345,14 +1348,14 @@ export class PortalManagementApi extends runtime.BaseAPI {
             query: queryParameters,
         }, initOverrides);
 
-        return new runtime.JSONApiResponse(response, (jsonValue) => OpenConsumerInfoDTOFromJSON(jsonValue));
+        return new runtime.JSONApiResponse(response, (jsonValue) => OpenConsumerTokenDTOFromJSON(jsonValue));
     }
 
     /**
      * GET /openapi/v1/consumer-tokens/by-appId
      * 按应用ID查询消费者Token(new added)
      */
-    async getConsumerTokenByAppId(requestParameters: GetConsumerTokenByAppIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<OpenConsumerInfoDTO> {
+    async getConsumerTokenByAppId(requestParameters: GetConsumerTokenByAppIdRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<OpenConsumerTokenDTO> {
         const response = await this.getConsumerTokenByAppIdRaw(requestParameters, initOverrides);
         return await response.value();
     }