+++ typescript/src/models/index.ts
@@ -12,6 +12,7 @@ export * from './OpenClusterNamespaceRoleUserDTO';
 export * from './OpenConsumerCreateRequestDTO';
 export * from './OpenConsumerInfoDTO';
 export * from './OpenConsumerSummaryDTO';
+export * from './OpenConsumerTokenDTO';
 export * from './OpenCreateAppDTO';
 export * from './OpenCreateNamespaceDTO';
 export * from './OpenEnvClusterDTO';